requirements.txt
python-docx==1.1.2
pyttsx3==2.98

